+++ .idea/.name
@@ -1,0 +1,1 @@
+Islami C42 G Sun Wed
