+++ app/src/main/java/com/route/islamic42gsunwed/activities/SplashIslamiActivity.kt
@@ -12,14 +12,10 @@ class SplashIslamiActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash_islami)
-        Handler(Looper.getMainLooper()).postDelayed(object : Runnable {
-            override fun run() {
-                navigateToHomeActivity()
-            }
-        }, 2000)
+        Handler(Looper.getMainLooper()).postDelayed({ navigateToHomeActivity() }, 2000)
     }
 
-    fun navigateToHomeActivity() {
+    private fun navigateToHomeActivity() {
         val intent = Intent(this, MainActivity::class.java)
         startActivity(intent)
         finish()
