--- app/src/main/java/com/route/islamic42gsunwed/fragments/TasbeehFragment.kt
@@ -1,26 +1,0 @@
-package com.route.islamic42gsunwed.fragments
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.route.islamic42gsunwed.databinding.FragmentQuranBinding
-import com.route.islamic42gsunwed.databinding.FragmentTasbeehBinding
-
-class TasbeehFragment : Fragment() {
-    private lateinit var binding: FragmentTasbeehBinding
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentTasbeehBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-
-}
